feat: implement seating and timetable solver services with associated REST resources, domain models, and validation tests.

+++ solver/src/main/java/com/icecream/seating/domain/SeatingSolution.java
@@ -81,11 +81,11 @@ public class SeatingSolution {
         this.score = score;
     }
 
-    public Integer getHardScore() {
+    public Long getHardScore() {
         return score == null ? null : score.hardScore();
     }
 
-    public Integer getSoftScore() {
+    public Long getSoftScore() {
         return score == null ? null : score.softScore();
     }
